[WPE] WPE Platform: move DRM implementation to create the WPEScreenSyncObserver from WPEScreenWayland to the base class https://bugs.webkit.org/show_bug.cgi?id=296472
Reviewed by Nikolas Zimmermann.

It could be reused for other platforms where DRM is available. Also add
some log debug messages to get information about the creation of the
WPEScreenSyncObserver.

Canonical link: https://commits.webkit.org/297856@main
