+++ Source/WebKit/WPEPlatform/wpe/WPEScreen.cpp
@@ -28,6 +28,16 @@
 
 #include <wtf/glib/WTFGType.h>
 
+#if USE(LIBDRM)
+#include "WPEScreenSyncObserverDRM.h"
+#include <errno.h>
+#include <fcntl.h>
+#include <optional>
+#include <wtf/glib/GRefPtr.h>
+#include <xf86drm.h>
+#include <xf86drmMode.h>
+#endif
+
 /**
  * WPEScreen:
  *
@@ -42,6 +52,9 @@ struct _WPEScreenPrivate {
     int physicalHeight { -1 };
     gdouble scale { 1 };
     int refreshRate { -1 };
+#if USE(LIBDRM)
+    GRefPtr<WPEScreenSyncObserver> syncObserver;
+#endif
 };
 
 WEBKIT_DEFINE_ABSTRACT_TYPE(WPEScreen, wpe_screen, G_TYPE_OBJECT)
@@ -137,6 +150,93 @@ static void wpeScreenGetProperty(GObject* object, guint propId, GValue* value, G
         G_OBJECT_WARN_INVALID_PROPERTY_ID(object, propId, paramSpec);
     }
 }
+
+static void wpeScreenInvalidate(WPEScreen* screen)
+{
+#if USE(LIBDRM)
+    screen->priv->syncObserver = nullptr;
+#endif
+}
+
+#if USE(LIBDRM)
+static std::optional<uint32_t> findCrtc(WPEScreen* screen, int fd)
+{
+    drmModeRes* resources = drmModeGetResources(fd);
+    if (!resources)
+        return std::nullopt;
+
+    std::optional<uint32_t> crtcIndex;
+    uint32_t widthMM = wpe_screen_get_physical_width(screen);
+    uint32_t heightMM = wpe_screen_get_physical_height(screen);
+    for (int i = 0; i < resources->count_connectors && !crtcIndex; ++i) {
+        auto* connector = drmModeGetConnector(fd, resources->connectors[i]);
+        if (!connector)
+            continue;
+
+        if (connector->connection != DRM_MODE_CONNECTED || !connector->encoder_id || !connector->count_modes) {
+            drmModeFreeConnector(connector);
+            continue;
+        }
+
+        if (widthMM != connector->mmWidth || heightMM != connector->mmHeight) {
+            drmModeFreeConnector(connector);
+            continue;
+        }
+
+        // FIXME: if there are multiple connectors matching the size, check other properties.
+        if (drmModeEncoder* encoder = drmModeGetEncoder(fd, connector->encoder_id)) {
+            for (int i = 0; i < resources->count_crtcs; ++i) {
+                if (resources->crtcs[i] == encoder->crtc_id) {
+                    crtcIndex = i;
+                    break;
+                }
+            }
+            drmModeFreeEncoder(encoder);
+        }
+        drmModeFreeConnector(connector);
+    }
+    drmModeFreeResources(resources);
+
+    return crtcIndex;
+}
+
+static void wpeScreenTryEnsureSyncObserver(WPEScreen* screen)
+{
+    drmDevicePtr devices[64];
+    const int devicesNum = drmGetDevices2(0, devices, std::size(devices));
+    if (devicesNum <= 0)
+        return;
+
+    for (int i = 0; i < devicesNum; i++) {
+        if (!(devices[i]->available_nodes & (1 << DRM_NODE_PRIMARY)))
+            continue;
+        auto fd = UnixFileDescriptor { open(devices[i]->nodes[DRM_NODE_PRIMARY], O_RDWR | O_CLOEXEC), UnixFileDescriptor::Adopt };
+        if (!fd)
+            continue;
+
+        if (auto crtcIndex = findCrtc(WPE_SCREEN(screen), fd.value())) {
+            screen->priv->syncObserver = adoptGRef(wpeScreenSyncObserverDRMCreate(WTFMove(fd), *crtcIndex));
+            if (screen->priv->syncObserver) {
+                g_debug("WPEScreen %u: Created WPEScreenSyncObserverDRM for device %s with CRTC index %u", screen->priv->id, devices[i]->nodes[DRM_NODE_PRIMARY], *crtcIndex);
+                break;
+            }
+        } else
+            g_debug("WPEScreen %u: Failed to find a CRTC for device %s", screen->priv->id, devices[i]->nodes[DRM_NODE_PRIMARY]);
+    }
+    drmFreeDevices(devices, devicesNum);
+
+    if (!screen->priv->syncObserver)
+        g_debug("WPEScreen %u: Could not create a WPEScreenSyncObserverDRM", screen->priv->id);
+}
+
+static WPEScreenSyncObserver* wpeScreenGetSyncObserver(WPEScreen* screen)
+{
+    auto* priv = screen->priv;
+    if (!priv->syncObserver)
+        wpeScreenTryEnsureSyncObserver(screen);
+    return priv->syncObserver.get();
+}
+#endif
 
 static void wpe_screen_class_init(WPEScreenClass* screenClass)
 {
@@ -259,6 +359,11 @@ static void wpe_screen_class_init(WPEScreenClass* screenClass)
             WEBKIT_PARAM_READWRITE);
 
     g_object_class_install_properties(objectClass, N_PROPERTIES, sObjProperties.data());
+
+    screenClass->invalidate = wpeScreenInvalidate;
+#if USE(LIBDRM)
+    screenClass->get_sync_observer = wpeScreenGetSyncObserver;
+#endif
 }
 
 /**
@@ -289,9 +394,7 @@ void wpe_screen_invalidate(WPEScreen* screen)
 {
     g_return_if_fail(WPE_IS_SCREEN(screen));
 
-    auto* screenClass = WPE_SCREEN_GET_CLASS(screen);
-    if (screenClass->invalidate)
-        screenClass->invalidate(screen);
+    WPE_SCREEN_GET_CLASS(screen)->invalidate(screen);
 }
 
 /**
